Major change to OpenGLSystemFactory
OpenGLSystem::Factory now creates components of type GLSprite. However
currently all they do is render a triangle.

Also a shader is now created in the component creation. This is horrible
placement and needs to be fixed.

Removed some redundant code and useless code.

Fixed rendering window height to be an actual value instead of 0
resulting in nothing being rendered.