Made user of properties in OpenGLSystem::Factory
OpenGLSystem::Factory for GLIcoSphere now uses the properties to set a
uniform scale and translation x, y, z.

Final fix for casting in OpenGLSystem::Update. Now it checks for a
dynamic_cast and throws std::bad_cast with a blank catch to allow the
next component type to check the cast.