First Implementation
Attempt at implementation of an ISystem class, that handles the
IComponent objects and the EntityID to component mapping inside the
System classes.

It is being used in the SimplePhysics and OpenGLSystem class, as well as
implementing SET_COMPONENT_ID(ComponentID) define to be able to set the
Component IDs for the different Components.