Adding WorldPosition component class.
WorldPosition stores the positions of the model in an efficient way
(contiguous elements for SIMD paradigm, and map access for finding
elements).

It proposes functions to access its data, and functions to give arrays
of translated model.

It can give a bitset of positions that are in the distance view of the
origin.

WorldPosition will be used in WorldObjectSystem.