Added in the ability to toggle mouse lock mode
By calling enabling mouse lock you are disabling the cursor.

Movement is still sent based on cursor movement from the center of the window.

There currently isn't a way to have the normal cursor with GLFW when you want to restrict the cursor to the current window.
A custom cursor could be draw at the virtual mouse coordinates.