Added mouse look tracking to FPSCamera
Now when the right mouse button is clicked mouse look mode is enabled.

Also moved the dispatch for mouse move events to before the position is stored.
This was causing movement deltas of 0 to be set because the old position was getting set to the new one before the dispatch.