Added the ability to request focus lock for keyboard events
Keyboard event handlers have a new system member that stores a pointer to the KeyboardInputSystem it is registered to.
By using this pointer handlers can request and release a focus lock.
In addition there is a new callback that is used when keyboard focus has been lost via a lock or window focus loss.

Also WebGUIView now checks for focus for keyboard events as well as the character events it did before.