First step of transformation of IMoverComponent into a real component in ECS.
In this step, IGLComponent is no more specific to an entity, but is a
global storage for a list of entities.

This implies:
- the creation of maps of data, using entity id as key.
- the addition of the entity id as parameter for each function
- the creation of entity management function (Add, Remove).

Children of IMoverComponent must use explicitly their own entity id to
call inherited functions as a temporary workaround.

The systems are not yet aware of the existence of the component.

Inheritance has not been broken yet to allow access to functions.