Second step of transformation of IMoverComponent into a real component in ECS.
Note that we will focus on the creation of the component that
interpolate a movement.

In this step, IGLComponent has now methods that were before in its child
BulletMover, and has taken some data specific to interpolation.
moreover, inheritance is now broken.

This implies:
- the creation of maps of the data taken from the child, using entity id
as key,
- The move of some code from the child to the component, and the
replacement by call to the static methods,
- The child does no more inherit from the component,
- the use of static functions everywhere for convenience: we do not have
yet a way to access to functions in the system.

The systems are not yet aware of the existence of the component.