GLTransform of IGLView now points to the one in ControllableMove that was created for the mover.
It was previously the contrary: IGLView created a GLTransform and made
the mover point on it.

This is more logical since the mover has a GLTransform because it is
controllable. It is not necessarily a viewer (for instance in a server
there are no viewers but there are remote movers).