Bring the CMakeLists.txt inline for all platforms
There is still a section for WIN32 to allow local placement of libraries and includes.
This section however only now adds extra search paths and doesn't hard code the libraries.

Also had to change FIndSOIL to search for a SOIL/ folder as well as include/SOIL/ and just inlucde/.
The issue is that SOIL is sometimes in include, include/SOIL, or in the case of the extra win32 include
paths in include/SOIL, but the search path starts in include so it just finds SOIL/.