Create an EntityManager, IEntity interface and the FeatureID
EntityManager is in charge of requesting addition of features, i.e a
predefined set of components, to ECS. A vector of objects representing
components are returned by ECS.

EntityManager will fill the IEntity component map with objects returned.
These objects wrap a weak_ptr on the primitive data of the component.

Features are the new name of the metacomponents such as
PhysicalWorldLocation or ControllableMove. They contain each several
components that share the same lifecycle and are strongly dependant.