Removed a redundant storing of transforms in PhysicalWorldLocation
This extra storing was causing a heap corruption due to creating the GLTransform on the stack,
emplacing it in the shared_ptr map, and then when that map was dtor it would call delete on
stack memory causing heap corruption.

I removed the duplicate map and just created the GLTransform in a shared_ptr to begin with
and store this in the transform_ptr_map directly.