Added ResourceSystem
To make an item a resource it my have a static TypeID member and provide a
bool Load(std::string) method. Then simply tell the ResourceSystem to CreateResourceLoader<MYRESOURCE>()
and you get then get resources of that type.

Texture now generates an ID in the constructor.

Several spots are using  __declspec(selectany) for static class members to create them in the header. To quote the gcc manual
"The selectany attribute is only available on Microsoft Windows targets. You can use __declspec (selectany) as a synonym for
 __attribute__ ((selectany)) for compatibility with other compilers. "
So it should be safely ignored unless under windows.